check if item to remove is a tuple or list

+++ physionet-django/project/utility.py
@@ -157,19 +157,21 @@ def remove_items(items, ignore_missing=True):
       items: Is a list of files or directories to delete.
       ignore_missing: Flag to ignore missing files or directories.
     """
-    while items:
-        try:
-            item = items.pop()
-            os.unlink(item)
-        except FileNotFoundError:
-            if not ignore_missing:
-                raise
-        except OSError as e:
-            if e.errno not in (errno.EISDIR, errno.EPERM):
-                raise
-            shutil.rmtree(item)
-        except (TypeError, AttributeError):
-            raise
+    if isinstance(items,(tuple,list)):
+        for item in items:
+            try:
+                os.unlink(item)
+            except FileNotFoundError:
+                if not ignore_missing:
+                    raise
+            except OSError as e:
+                if e.errno not in (errno.EISDIR, errno.EPERM):
+                    raise
+                shutil.rmtree(item)
+    else:
+        LOGGER.info("Non list/tuple entered in remove items. The 'item' entered \
+         was: {0}".format(items))
+        raise TypeError
 
 def clear_directory(directory):
     """
